Move iframe class to top level.
