+++ src/default-theme/help.css
@@ -11,24 +11,3 @@
 .jp-Help > iframe {
   border: none;
 }
-
-
-/*
-When drag events occur, `p-mod-override-cursor` is added to the body.
-Because iframes steal all cursor events, the following two rules are necessary
-to suppress pointer events while resize drags are occuring. There may be a
-better solution to this problem.
-*/
-body.p-mod-override-cursor .jp-Help {
-  position: relative;
-}
-
-body.p-mod-override-cursor .jp-Help:before {
-  content: '';
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background: transparent;
-}
