+++ src/index.css
@@ -5,5 +5,6 @@
 |----------------------------------------------------------------------------*/
 @import './dialog/index.css';
 @import './filebrowser/index.css';
+@import './iframe/index.css';
 @import './notebook/completion/index.css'
 @import './terminal/index.css';
